+++ Tests/PlayerTests/PlayerSeekTests.swift
@@ -12,49 +12,6 @@ import Nimble
 import XCTest
 
 final class PlayerSeekTests: XCTestCase {
-    func testCannotSeekWithEmptyPlayer() {
-        let player = Player()
-        expect(player.canSeek(to: .zero)).to(beFalse())
-    }
-
-    func testCanSeekInTimeRange() {
-        let player = Player(item: .simple(url: Stream.onDemand.url))
-        expect(player.streamType).toEventually(equal(.onDemand))
-        expect(player.canSeek(to: CMTimeMultiplyByFloat64(Stream.onDemand.duration, multiplier: 0.5))).to(beTrue())
-    }
-
-    func testCannotSeekInEmptyTimeRange() {
-        let player = Player(item: .simple(url: Stream.live.url))
-        expect(player.streamType).toEventually(equal(.live))
-        expect(player.canSeek(to: .zero)).to(beFalse())
-    }
-
-    func testCanSeekToTimeRangeStart() {
-        let player = Player(item: .simple(url: Stream.onDemand.url))
-        expect(player.streamType).toEventually(equal(.onDemand))
-        expect(player.canSeek(to: player.timeRange.start)).to(beTrue())
-    }
-
-    func testCanSeekToTimeRangeEnd() {
-        let player = Player(item: .simple(url: Stream.onDemand.url))
-        expect(player.streamType).toEventually(equal(.onDemand))
-        expect(player.canSeek(to: player.timeRange.end)).to(beTrue())
-    }
-
-    func testCannotSeekBeforeTimeRangeStart() {
-        let player = Player(item: .simple(url: Stream.onDemand.url))
-        expect(player.streamType).toEventually(equal(.onDemand))
-        expect(player.canSeek(to: CMTime(value: -10, timescale: 1))).to(beFalse())
-    }
-
-    func testCannotSeekAfterTimeRangeEnd() {
-        let player = Player(item: .simple(url: Stream.onDemand.url))
-        expect(player.streamType).toEventually(equal(.onDemand))
-        expect(player.canSeek(to: player.timeRange.end + CMTime(value: 1, timescale: 1))).to(beFalse())
-    }
-
-    // ---
-
     func testSeekWithEmptyPlayer() {
         let player = Player()
         waitUntil { done in
