+++ Tests/PlayerTests/PlayerSkipBackwardChecksTests.swift
@@ -1,0 +1,37 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+@testable import Player
+
+import Circumspect
+import CoreMedia
+import Nimble
+import XCTest
+
+final class PlayerSkipBackwardChecksTests: XCTestCase {
+    func testCannotSkipWhenEmpty() {
+        let player = Player()
+        expect(player.canSkipBackward()).to(beFalse())
+    }
+
+    func testCanSkipForOnDemand() {
+        let player = Player(item: .simple(url: Stream.onDemand.url))
+        expect(player.streamType).toEventually(equal(.onDemand))
+        expect(player.canSkipBackward()).to(beFalse())
+    }
+
+    func testCanSkipForLive() {
+        let player = Player(item: .simple(url: Stream.live.url))
+        expect(player.streamType).toEventually(equal(.live))
+        expect(player.canSkipBackward()).to(beFalse())
+    }
+
+    func testCanSkipForDvr() {
+        let player = Player(item: .simple(url: Stream.dvr.url))
+        expect(player.streamType).toEventually(equal(.dvr))
+        expect(player.canSkipBackward()).to(beTrue())
+    }
+}
