+++ Tests/PlayerTests/PlayerSkipBackwardTests.swift
@@ -12,29 +12,6 @@ import Nimble
 import XCTest
 
 final class PlayerSkipBackwardTests: XCTestCase {
-    func testCannotSkipWhenEmpty() {
-        let player = Player()
-        expect(player.canSkipBackward()).to(beFalse())
-    }
-
-    func testCanSkipForOnDemand() {
-        let player = Player(item: .simple(url: Stream.onDemand.url))
-        expect(player.streamType).toEventually(equal(.onDemand))
-        expect(player.canSkipBackward()).to(beFalse())
-    }
-
-    func testCanSkipForLive() {
-        let player = Player(item: .simple(url: Stream.live.url))
-        expect(player.streamType).toEventually(equal(.live))
-        expect(player.canSkipBackward()).to(beFalse())
-    }
-
-    func testCanSkipForDvr() {
-        let player = Player(item: .simple(url: Stream.dvr.url))
-        expect(player.streamType).toEventually(equal(.dvr))
-        expect(player.canSkipBackward()).to(beTrue())
-    }
-
     func testSkipForOnDemand() {
         let player = Player(item: .simple(url: Stream.onDemand.url))
         expect(player.streamType).toEventually(equal(.onDemand))
